Advanced/requirements.txt
Flask==1.1.2
google-cloud-datastore==1.12.0
json2html==1.3.0